Fix clone session programming (#59)
The move from BFRT to TDI removed the "$" prefix from certain table names, actions and fields for most fixed tables. Clone is one of them. A future change like this is required for `pre` and other components.

The clone session programming in the current form is Tofino specific, although it sits in the common area. If and when DPDK or MEV support clone sessions, this piece of code would require refactoring into target specific code.

The current set of changes are minimally modifying the string values to make the code work with TDI tofino. The $ prefixes are a BFRT naming convention done away with in TDI.

Signed-off-by: Ravi Vantipalli <ravi.vantipalli@intel.com>