Fix bazel test error on CI server (#39)
Problem: The CI server fails with the Python exception
`AttributeError: module 'collections' has no attribute 'MutableMapping'`
when running Stratum unit tests under Bazel.

MutableMapping was moved from 'collections' to 'collections.abc' in
Python 3.3. Support for the old import was deprecated, and it was
announced that it would be removed after Python 3.9. The CI server
is running Python 3.10.

On examination, we found that:
1) The code containing the old import is in rules_python.
2) Our version of Stratum uses (pre)release 0.0.1 of rules_python.
   The latest release of rules_python is 0.21.0.
3) The upstream version of Stratum has eliminated rules_python
   entirely.

We have addressed the issue by merging changes from upstream
to eliminate rules_python from our repository as well.

Signed-off-by: Derek G Foster <derek.foster@intel.com>