Implement TargetOptions class
Replace the compile-time conditionals in ConfigMonitoringService and
P4Service with run-time tests on a parameter object (TargetOptions)
that is specified when the service is instantiated.

Use TargetOptions to control whether:

1) ConfigMonitoringService requires that the chassis config file
   not be a symlink. (enabled for security)
2) ConfigMonitoringService pushes an updated chassis configuration.
   (disabled for DPDK)
3) P4Service allows a configured pipeline to be overwritten.
   (disabled for DPDK)

Update DpdkMain to pass a TargetOptions object to DpdkHal, which in
turn passes it to ConfigMonitoringService and P4Service. Modify
constructors and factory methods to accept the TargetOption object
while maintaining backward compatibility with existing code.

The secureChassisConfig option is currently disabled, to support
unit testing. Further work is needed to allow this option to be
enabled by default for TDI targets and disabled by default for
non-TDI targets and unit tests.

Signed-off-by: Derek G Foster <derek.foster@intel.com>