Remove MVLUT specific changes from Startum
Because we are updating the p4runtime submodule to pick from
latest p4runtime and ignoring custom changes made to
p4runtime.proto and p4info.proto to support Match Value Lookup,
we will not have the definitions of MvLUT in the autogenerated
p4runtime.pb.h/.cc files. This causes a build failure.

These changes are made to remove all LUT specific changes in stratum.
WriteExtern API is kept intact so that future Meter changes can go in
there.

Signed-off-by: Sandeep N <sandeep.nagapattinam@intel.com>