[sync] Rename `unit` to `device` in TDI code
Tracks upstream changes to the Barefoot code made in Stratum PR #732.

`unit` is a Broadcom-specific term and has no established meaning in the
Tofino/TDI world, where `device` is used for the same purpose. To reduce
cognitive load and prevent unnecessary confusion, we replace the term.

This PR does contain any functional changes.

Signed-off-by: Derek G Foster <derek.foster@intel.com>