Cherry pick from 1.6 (#234)
Some of the previous log level noise reduction is being undone
here. Since we started reducing the amount of logging, error
conditions have become too generic, saying "one or more
read/write operations failed" with no information on what the
underlying problem could be. (Was it a duplicate entry, or
table_not_found, or some other internal error? It's hard to tell).
Restoring some of the previous error logging gives the user more
detailed information on the nature of the problem.

Signed-off-by: Sabeel Ansari <sabeel.ansari@intel.com>