Introduce ExternManager class architecture
This CL implements an architecture that can be used to isolate
code that depends on the ES2K-specific IDPF extensions to
P4Runtime from the rest of the code.

As things stand right now, we cannot upstream our changes to
P4InfoManager because they break non-IPDK Stratum platforms
(barefoot, bcm, fmv2, etc.) unless they are built against our
custom P4Runtime. The new architecture adds an interface that
allows platform-specific extensions to P4InfoManager to be
incorporated at build time.

The architecture also provides a mechanism that supports custom
extensions to TdiTableManager to be incorporated at build time.

The IDPF-dependent code goes into the extensions, which are
compiled and linked in as part of the ES2K build.

This CL adds the Es2kExternManager class and its relatives,
together with a unit test and Bazel updates to support the
new components.

It does NOT include the P4InfoManager and TdiTableManager changes
that remove the existing IDPF support and replace it with the
ExternManager code. The new files will be passive additions to
the codebase until this is done (in a subsequent commit).

Signed-off-by: Derek Foster <derek.foster@intel.com>