Implement P4InfoManager extension classes
We cannot upstream our modifications to P4InfoManager
in their current form because they are ES2K-specific and
make the module (which is common to all Stratum platforms)
depend directly on a non-standard version of P4Runtime.

This CL defines a family of classes that allow a platform to
supply an extension that implements custom P4Extern processing
without requiring changes to P4InfoManager itself.

It also includes an implementation of an extension for ES2K,
and a unit test for the ES2K extension.

A subsequent commit will replace the ES2K-specific code in
P4InfoManager with support for the extension mechanism, and
it will modify TdiTableManager to supply the extension.

Signed-off-by: Derek Foster <derek.foster@intel.com>